requirement.txt
pyTableMaker
discord.py
chatterbot==0.8.7
chatterbot-corpus
mcstatus
flask
python-dotenv
numpy
psutil
duckduckgo3